Fix eventid int overflow after bigint migration
events.eventid and related FK columns (svclosteventid, svcregainedeventid,
notifications.eventid, event_parameters.eventid, alarms.lasteventid) were
migrated to bigint in 34.0.0, but several legacy JDBC and servlet paths
were not updated, causing crashes and silent data corruption once the
sequence exceeds Integer.MAX_VALUE (2147483647).

- EventFactory.rs2Events: rs.getInt("eventID") -> rs.getLong (crash on node page)
- NotificationManager.createNotification: setInt/parseInt -> setLong/parseLong for
  eventid INSERT parameter; setNull type INTEGER -> BIGINT
- NotificationManager.acknowledgeNotificationsForEvent: int eventID -> long, getLong
- NotificationManager.getEvent: widen parameter int -> long
- BroadcastEventProcessor: Integer.parseInt -> Long.parseLong for getEvent() call
- OutageModel.rs2Outages: rs.getInt("svclosteventid") -> rs.getLong
- NoticeFactory.rs2Notices: Integer.valueOf(rs.getInt("eventid")) -> rs.getLong
- NotificationModel.rs2NotifyBean: rs.getInt(EVENTID) -> rs.getLong
- EventQueryServlet: safeParseInt -> safeParseLong for eventid filter
- EventUtil.getFilter: safeParseInt -> safeParseLong for EventIdFilter

+++ opennms-webapp/src/main/java/org/opennms/web/outage/OutageModel.java
@@ -223,7 +223,7 @@ public abstract class OutageModel {
             }
 
             if (includesNotifications) {
-                long serviceLostEventId = rs.getInt("svclosteventid");
+                long serviceLostEventId = rs.getLong("svclosteventid");
                 if (!rs.wasNull()) {
                     outage.lostServiceEventId = serviceLostEventId;
                 }
